Do not update when a special buffer is selected
And “special” here, means anything else that a normal buffer (a file).

Additionaly make sure we reset the alternate window as well so we can keep using that as expected.

For details why this fix is needed, please see this issue https://github.com/ctrlpvim/ctrlp.vim/issues/235 and the comments (mainly the last 5 or so) of this PR https://github.com/ctrlpvim/ctrlp.vim/pull/303