Adding "CwdOnly" option to limit the scope of git status
For repositories with a very large number of files (>100k), the `git
status` command can take a very long time. A common practice in these
sorts of repositories is to not open the editor at the root of the
repository, but rather within the particular subcomponent being modified
at the time. By adding this scoping option, the performance of NERDTree
is vastly improved in these situations.